+++ benchmarks/engine_benchmark.cpp
@@ -49,7 +49,7 @@ void parse_cmdline(int argc, char *argv[]) {
   aos("enable-metrics", po::value<bool>(&metrics_enable)->default_value(metrics_enable));
   aos("read-benchmark-runs", po::value<size_t>(&read_benchmark_runs)->default_value(read_benchmark_runs));
   aos("strategy", po::value<STRATEGY>(&strategy)->default_value(strategy), "Write strategy");
-  aos("memory-limit", po::value<size_t>(&memory_limit)->default_value(memory_limit), "allocation area limit when strategy=MEMORY");;
+  aos("memory-limit", po::value<size_t>(&memory_limit)->default_value(memory_limit), "allocation area limit  in megabytes when strategy=MEMORY");
 
   po::variables_map vm;
   try {